Refactoring structure to isolate todo components
