+++ App.tsx
@@ -1,19 +1,15 @@
-import { StyleSheet, View, StatusBar } from 'react-native';
+import React, { StyleSheet, View, StatusBar } from 'react-native';
 import { TodoProvider } from './src/contexts/TodoContext';
-import Header from './src/components/Header';
-import TodoList from './src/components/TodoList';
-import AddItem from './src/components/AddItem';
+import Todos from './src/components/Todos/Todos'
 
 export default function App() {
   return (
-    <TodoProvider>
-      <View style={styles.container}>
-        <StatusBar barStyle="dark-content" />
-        <Header />
-        <TodoList />
-        <AddItem />
-      </View>
-    </TodoProvider>
+    <View style={styles.container}>
+      <StatusBar barStyle="dark-content" />
+      <TodoProvider>
+        <Todos />
+      </TodoProvider>
+    </View>
   );
 }
